# Apply list view optimizations to detail view