job detail view, `/api/v2/jobs/N/`

```
before:
X-API-Query-Count: 44
X-API-Query-Time: 0.033s
X-API-Time: 1.241s

after:
X-API-Query-Count: 15
X-API-Query-Time: 0.017s
X-API-Time: 0.695s

after, once it got caching going:

X-API-Query-Count: 10
X-API-Query-Time: 0.014s
X-API-Time: 0.226s
```

job template detail view:

```
before:
X-API-Query-Count: 31
X-API-Query-Time: 0.025s
X-API-Time: 0.732s

after:
X-API-Query-Count: 10
X-API-Query-Time: 0.014s
X-API-Time: 0.228s
```

The mechanism here: `get_object` does `get_object_or_404`, but it feed the queryset into this. This PR adds optimization to the model's default manager's queryset.