@AlanCoding 2017-12-11
@ryanpetrello the increase in detail view queries is fairly unique to `devel` and JTs / jobs because of the combination of multi-credential with jumping 2 relationships over to credential_type. Quite a large number of references are made to credential `kind` and things like that, even from the job / JT, plus things like `vault_credential`. This is also why I always wound up looping like `[cred for cred in obj.credentials.all() if cred.kind == 'ssh']` instead of filtering. The prefetched cache is only hit using the `.all()` query.