@ryanpetrello 2017-12-13
Other than my `KeyError` comment, this looks good.